Revert "Patch:Linux 3.4.11-3.4.105"
I don't want to pay any tribute or shit to anybody that thinks low of me..thats it..I'm reverting this as my kernel is still best even without it and the fact that HTC or Google could have updated their own kernel bases to the latest and cutting edge Linux version..but they did not..we are running a smart phone and not a desktop here..

This reverts commit 761da6a2a8789ced1a0b72b5315475b2da6ebaff.

+++ include/linux/msg.h
@@ -59,14 +59,13 @@ struct msginfo {
 #ifdef __KERNEL__
 #include <linux/list.h>
 
-/* one msg_msg structure for each message */
 struct msg_msg {
-	struct list_head m_list;
-	long m_type;
-	size_t m_ts;		/* message text size */         
+	struct list_head m_list; 
+	long  m_type;          
+	int m_ts;           
 	struct msg_msgseg* next;
 	void *security;
-	/* the actual message follows immediately */
+	
 };
 
 struct msg_queue {
